+++ src/compiler/sema_types.c
@@ -262,6 +262,7 @@ static bool sema_resolve_type_identifier(SemaContext *context, TypeInfo *type_in
 		case DECL_DECLARRAY:
 		case DECL_BODYPARAM:
 		case DECL_CT_INCLUDE:
+		case DECL_CT_EXPAND:
 		case DECL_CT_EXEC:
 		case DECL_GLOBALS:
 			UNREACHABLE